Improve Trace Event
